Upgraded to version 2.2. Upgraded SDK requirements to Cupcake. Updated README file so the instructions work with the latest SDK.
Added support for 12.2 kbps 3gpp/amr files, the kind recorded natively by Android 1.6 "Donut" and beyond.

Added a random but unique id to anonymous metadata uploaded to Ringdroid.com so that we can count each song from each user only once.

Improved accuracy of the start and end time of MP3 files, so that what's heard while editing is exactly the same as what's heard when the clipped file is extracted.  The same technique isn't possible for other file formats.

Fixed layout and drawing bugs on high-resolution screens.

Fixed bug where the waveform would not auto-scroll to the start or end markers when both were offscreen.

Fixed crash that would happen if the screen tried to redraw while the sound file was still being loaded.